+++ test/CsrfMiddlewareTest.php
@@ -1,17 +1,19 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-csrf for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-csrf/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace ZendTest\Expressive\Csrf;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use Zend\Expressive\Csrf\CsrfGuardFactoryInterface;
 use Zend\Expressive\Csrf\CsrfGuardInterface;
 use Zend\Expressive\Csrf\CsrfMiddleware;
@@ -63,9 +65,9 @@ class CsrfMiddlewareTest extends TestCase
         $this->guardFactory->createGuardFromRequest($request->reveal())->willReturn($guard);
         $request->withAttribute($attributeKey, $guard)->will([$request, 'reveal']);
 
-        $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process(Argument::that([$request, 'reveal']))->willReturn($response);
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->handle(Argument::that([$request, 'reveal']))->willReturn($response);
 
-        $this->assertSame($response, $middleware->process($request->reveal(), $delegate->reveal()));
+        $this->assertSame($response, $middleware->process($request->reveal(), $handler->reveal()));
     }
 }