+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.0.0 - 2018-03-15
+
+### Added
+
+- [#5](https://github.com/zendframework/zend-expressive-csrf/pull/5) adds
+  support for PSR-15.
+
+### Changed
+
+- Updates minimum supported versions of both zend-expressive-session and
+  zend-expressive-flash to 1.0.0.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- [#5](https://github.com/zendframework/zend-expressive-csrf/pull/5) and
+  [#3](https://github.com/zendframework/zend-expressive-csrf/pull/3) remove
+  support for http-interop/http-middleware and
+  http-interop/http-server-middleware.
+
+### Fixed
+
+- Nothing.
+
 ## 0.1.1 - 2017-10-11
 
 ### Added